Added vue router, and routed TorrentPlayer, and settings page
