+++ package.json
@@ -61,6 +61,7 @@
     "vue": "^2.3.3",
     "vue-electron": "^1.0.6",
     "vue-material": "^0.7.4",
+    "vue-router": "^3.0.1",
     "vuex": "^2.4.0",
     "webtorrent": "^0.98.19"
   },
