+++ src/renderer/App.vue
@@ -1,29 +1,11 @@
-<template>
-  <div id="app">
-    <!-- <landing-page></landing-page> -->
-    <top-bar></top-bar>
-    <playlist></playlist>
-    <player></player>
-  </div>
+<template lang="html">
+  <router-view></router-view>
 </template>
 
 <script>
-  import LandingPage from '@/components/LandingPage'
-  import SongList from '@/components/SongList.vue'
-  import Player from '@/components/Player.vue'
-  import topBar from '@/components/Header.vue'
-
-  export default {
-    name: 'torrent-player',
-    components: {
-      LandingPage,
-      'playlist': SongList,
-      Player,
-      topBar
-    }
-  }
+export default {
+}
 </script>
 
-<style>
-  /* CSS */
+<style lang="css">
 </style>
